add std::initializer_list support for boost::container::vector
